Fix for task id wrap round
git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@162 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/tags/1.2.8/src/mydns/task.c
@@ -27,7 +27,7 @@
 extern void server_status(void);
 extern void named_cleanup(int);
 
-static uint16_t		internal_id = 0;
+static uint32_t		internal_id = 0;
 static uint8_t		*taskvec = NULL;
 static int32_t		active_tasks = 0;
